première versinde l'api

+++ src/main/java/com/testoc/apitest/model/Hospital.java
@@ -16,7 +16,7 @@ public class Hospital {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  public Long id;
+  public int id;
 
   @Column(name="name")
   public String Name;
